+++ src/conformly/_internal/api/case.py
@@ -59,6 +59,12 @@ def case(
             - PathSelector DSL:
                 path("user.email").violate(V.TOO_SHORT)
 
+        overrides:
+            Optional list of `PathSelector` to set field values.
+            - Applied after generation
+            - Override generated values
+            - With valid=False: act as defaults (ignored if field is violated)
+
         allow_type_mismatch:
             If True fields could be type mismatched.
             Availiable only when valid=False.