+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+- Now `path` uses proxy-object instead of ast parsing
+
 
 ## [0.6.0] - 2026.06.13